requirements.txt
Pillow
matplotlib
numpy
opencv-contrib-python
opencv-python-headless
selectivesearch
xmltodict
streamlit
torch
torchvision